+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
-- N/A
+## [Unreleased] - TBD
+### Fixed
+- Bug where basic lands never matched the quantity that the deck requires.
 
 ## [0.1.6] - 9-16-2019
 ### Fixed